v3: reduce self-host peak memory another ~10%, performance-neutral
Benchmark: `cd vlib/v3 && ../../v -gc none -prealloc -prod -o v3 v3.v &&
./v3 -nocache -building-v -o v4 v3.v`, 3 interleaved pairs on an idle
machine: peak footprint 726/740/727 MB -> 652/655/655 MB (-10.5%), max RSS
734/748/735 -> 662/663/663 (-10.3%), instructions retired 45.71 G -> 45.76 G
(noise). Codegen is byte-identical to the baseline compiler on the same
sources, also under -d prealloc_no_recycle, and the compiler reproduces
itself (gen2).

- flat.Node 88 -> 80 bytes: the payload pointer becomes a u32 id into a
  process-wide chunked table (copying a node copies the id, so no merge path
  needs remapping), and the two rare bools become bits in a flags byte with
  accessor methods.
- Resolved fn-value names: the dense per-node array plus set bits (12 MB)
  held 81 entries after the check; the sparse map is now the only store,
  parallel-check forks get fresh maps and merge after the join.
- Shared transform helpers no longer clone the used_fns map each: they read
  the immutable root snapshot and record only their own discoveries, and the
  merge no longer re-inserts every entry of every copy. The two whole-set
  scans walk own, parent and root maps.
- Transform helper batches 8 -> 16, halving the batch arenas live at the
  transform peak; node module/file context caches u32 -> u16.
- cgen: const initializer lowering (12 MB of the type-declaration task's
  permanent base garbage) runs on a private worker inside a disposable arena;
  the emission order still comes from the master, whose dependency memos are
  warm, and only the text, the runtime-init queue and the batch side tables
  are published.
- Tests that hard-coded the removed fields or the old cache width are
  updated.

+++ vlib/v3/transform/monomorphize_test.v
@@ -4,16 +4,16 @@ import v3.flat
 import v3.types
 
 fn test_node_context_cache_growth_preserves_ids_and_initializes_new_slots() {
-	mut cache := []u32{}
+	mut cache := []u16{}
 	grow_node_context_cache(mut cache, 3, 8)
 	cache[0] = 7
 	cache[2] = 9
 	grow_node_context_cache(mut cache, 6, 8)
-	assert cache == [u32(7), 0, 9, 0, 0, 0]
+	assert cache == [u16(7), 0, 9, 0, 0, 0]
 	cache[5] = 11
 	grow_node_context_cache(mut cache, 12, 16)
-	assert cache[..6] == [u32(7), 0, 9, 0, 0, 11]
-	assert cache[6..] == [u32(0), 0, 0, 0, 0, 0]
+	assert cache[..6] == [u16(7), 0, 9, 0, 0, 11]
+	assert cache[6..] == [u16(0), 0, 0, 0, 0, 0]
 	grow_node_context_cache(mut cache, 2, 16)
 	assert cache.len == 12
 	assert cache[5] == 11
